requirements.txt
flask
python-telegram-bot
beautifulsoup4
html5lib